Remove duplicate BookmarkTransactionBuilder
