+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/logic/commands/bookmarkcommands/AddBookmarkExpenseCommandTest.java
@@ -14,10 +14,11 @@ import org.junit.jupiter.api.Test;
 import ay2021s1_cs2103_w16_3.finesse.logic.commands.CommandResult;
 import ay2021s1_cs2103_w16_3.finesse.logic.commands.bookmark.AddBookmarkExpenseCommand;
 import ay2021s1_cs2103_w16_3.finesse.logic.commands.stubs.ModelStub;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.bookmarkparsers.BookmarkTransactionBuilder;
 import ay2021s1_cs2103_w16_3.finesse.model.FinanceTracker;
 import ay2021s1_cs2103_w16_3.finesse.model.ReadOnlyFinanceTracker;
 import ay2021s1_cs2103_w16_3.finesse.model.bookmark.BookmarkExpense;
-import ay2021s1_cs2103_w16_3.finesse.testutil.BookmarkTransactionBuilder;
+
 
 public class AddBookmarkExpenseCommandTest {
