+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/model/bookmark/BookmarkIncomeTest.java
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import ay2021s1_cs2103_w16_3.finesse.testutil.BookmarkTransactionBuilder;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.bookmarkparsers.BookmarkTransactionBuilder;
 
 public class BookmarkIncomeTest {
